chore: Change Readme file and fix some visual bugs
