+++ lib/cjs/config/index.d.ts
@@ -1,18 +1,4 @@
-export interface Theme {
-    yearColor: string;
-    lineColor: string;
-    dotColor: string;
-    borderDotColor: string;
-    titleColor: string;
-    subtitleColor: string;
-    textColor: string;
-}
-export declare type DateFormat = 'only-number' | 'short' | 'with-weekday' | 'full';
-export interface DefaultValues {
-    theme?: Theme;
-    lang: string;
-    dateFormat: DateFormat;
-}
+import { DefaultValues } from '../interfaces';
 export declare const defaultValues: DefaultValues;
 export declare const mapText: {
     en: {
