+++ lib/cjs/context/config.context.d.ts
@@ -1,2 +1,2 @@
-import { DefaultValues } from '../config';
+import { DefaultValues } from '../interfaces';
 export declare const ConfigContext: import("react").Context<DefaultValues>;
