+++ lib/esm/styles/main.d.ts
@@ -1,4 +1,4 @@
-import { DateFormat } from '../config';
+import { DateFormat } from '../interfaces';
 export declare const TimelineWrapper: import("styled-components").StyledComponent<"div", any, {}, never>;
 export declare const TimelineWrapperInner: import("styled-components").StyledComponent<"div", any, {}, never>;
 export declare const ContainerWrapper: import("styled-components").StyledComponent<"section", any, {}, never>;
