+++ lib/esm/styles/main.js
@@ -18,16 +18,19 @@ var minWidth = function () { return ({
     short_zh: '115',
     'with-weekday_es': '125',
     'with-weekday_en': '125',
-    'with-weekday_de': '120',
-    'with-weekday_zh': '120',
-    full_es: '135',
-    full_en: '135',
-    full_de: '135',
-    full_zh: '135',
+    'with-weekday_de': '130',
+    'with-weekday_zh': '140',
+    full_es: '125',
+    full_en: '125',
+    full_de: '125',
+    full_zh: '125',
 }); };
-export var YearWrapper = styled.p(templateObject_5 || (templateObject_5 = __makeTemplateObject(["\n\talign-items: flex-end;\n\tcolor: ", ";\n\tdisplay: flex;\n\tflex-direction: column;\n\tfont-size: 14px;\n\tfont-weight: 400;\n\theight: max-content;\n\tmargin: 0;\n\tmax-width: 35%;\n\tmin-width: ", ";\n\ttext-align: end;\n"], ["\n\talign-items: flex-end;\n\tcolor: ", ";\n\tdisplay: flex;\n\tflex-direction: column;\n\tfont-size: 14px;\n\tfont-weight: 400;\n\theight: max-content;\n\tmargin: 0;\n\tmax-width: 35%;\n\tmin-width: ", ";\n\ttext-align: end;\n"])), function (_a) {
+export var YearWrapper = styled.p(templateObject_5 || (templateObject_5 = __makeTemplateObject(["\n\talign-items: flex-end;\n\tcolor: ", ";\n\tdisplay: flex;\n\tflex-direction: column;\n\tfont-size: 14px;\n\tfont-weight: 400;\n\theight: max-content;\n\tmargin: 0;\n\tmax-width: ", ";\n\tmin-width: ", ";\n\ttext-align: end;\n"], ["\n\talign-items: flex-end;\n\tcolor: ", ";\n\tdisplay: flex;\n\tflex-direction: column;\n\tfont-size: 14px;\n\tfont-weight: 400;\n\theight: max-content;\n\tmargin: 0;\n\tmax-width: ", ";\n\tmin-width: ", ";\n\ttext-align: end;\n"])), function (_a) {
     var theme = _a.theme;
     return theme.yearColor;
+}, function (_a) {
+    var format = _a.format, lang = _a.lang;
+    return "".concat(minWidth()["".concat(format, "_").concat(lang)], "px");
 }, function (_a) {
     var format = _a.format, lang = _a.lang;
     return "".concat(minWidth()["".concat(format, "_").concat(lang)], "px");
